+++ ECommerceBackend/src/main/java/com/practice/ecommerce/dto/OrderedProductsResponseDto.java
@@ -2,6 +2,7 @@ package com.practice.ecommerce.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -9,5 +10,5 @@ public class OrderedProductsResponseDto {
 
     private List<ProductDto> productDtoList;
 
-    private Long orderAmount;
+    private BigDecimal orderAmount;
 }
